requirements.txt
rsa~=4.9
mysql-connector~=2.2.9
pycryptodome~=3.20.0